+++ core/api_container/api_container_docker_consts/api_container_mountpoints/api_container_mountpoints.go
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package api_container_mountpoints
+
+const (
+	// The location where the suite execution Docker volume will be mounted on the API container
+	SuiteExecutionVolumeMountDirpath = "/suite-execution"
+)