+++ core/api_container/server/server_core.go
@@ -7,7 +7,7 @@ package server
 
 import (
 	"github.com/kurtosis-tech/kurtosis/api_container/api/bindings"
-	"github.com/kurtosis-tech/kurtosis/api_container/exit_codes"
+	"github.com/kurtosis-tech/kurtosis/api_container/api_container_docker_consts/api_container_exit_codes"
 	"google.golang.org/grpc"
 )
 
@@ -16,5 +16,5 @@ type ApiContainerServerCore interface {
 	GetSuiteAction() bindings.SuiteAction
 
 	// Creates a server using the given shutdownChan, and registers it with the given grpc server
-	CreateAndRegisterService(shutdownChan chan exit_codes.ApiContainerExitCode, grpcServer *grpc.Server) ApiContainerServerService
+	CreateAndRegisterService(shutdownChan chan api_container_exit_codes.ApiContainerExitCode, grpcServer *grpc.Server) ApiContainerServerService
 }