+++ core/api_container/server/suite_metadata_serialization/suite_metadata_serialization_server_core.go
@@ -7,7 +7,7 @@ package suite_metadata_serialization
 
 import (
 	"github.com/kurtosis-tech/kurtosis/api_container/api/bindings"
-	"github.com/kurtosis-tech/kurtosis/api_container/exit_codes"
+	"github.com/kurtosis-tech/kurtosis/api_container/api_container_docker_consts/api_container_exit_codes"
 	"github.com/kurtosis-tech/kurtosis/api_container/server"
 	"google.golang.org/grpc"
 )
@@ -24,7 +24,7 @@ func (core SuiteMetadataSerializationServerCore) GetSuiteAction() bindings.Suite
 	return bindings.SuiteAction_SERIALIZE_SUITE_METADATA
 }
 
-func (core SuiteMetadataSerializationServerCore) CreateAndRegisterService(shutdownChan chan exit_codes.ApiContainerExitCode, grpcServer *grpc.Server) server.ApiContainerServerService {
+func (core SuiteMetadataSerializationServerCore) CreateAndRegisterService(shutdownChan chan api_container_exit_codes.ApiContainerExitCode, grpcServer *grpc.Server) server.ApiContainerServerService {
 	service := newSuiteMetadataSerializationService(core.serializationOutputFilepath, shutdownChan)
 	bindings.RegisterSuiteMetadataSerializationServiceServer(grpcServer, service)
 	return service